HalfTensor

+++ PoetGen/trainer_torch.py
@@ -33,7 +33,7 @@ class Trainer:
                     inputs = inputs * mask.int()
                 
                 out = self.model(input_ids=inputs.to(self.device), labels=label.to(self.device), 
-                                 attention_mask=batch['attention'].to(self.device), vowel_count=(batch['nums'].type(torch.FloatTensor)).to(self.device))
+                                 attention_mask=batch['attention'].to(self.device), vowel_count=(batch['nums'].type(torch.HalfTensor)).to(self.device))
                 out['model_output'].loss.backward(retain_graph=True)             
                 if self.consistency_task:
                     out['vowel_regression_loss'].backward()
